requirements.txt
Flask==3.0.0
tensorflow==2.20.0
Pillow==10.0.0
numpy==1.26.0
opencv-python-headless==4.8.0.76
scipy
gunicorn==21.2.0
werkzeug==3.0.1
gdown==4.7.1